Use inherited Slideshow class
I inherit the Slideshow class in ChefShow but wasn't using it. This was
some left over implementation from before the refactor. Now correctly
use the super.

Added a update_progress method in ChefShow to override the one in
Slideshow so the caller does not need to pass all the arguments since
the class will know them.