Remove init()
Remove init() function from ChefShow class and move the code to
__init__.

The reason for the code being separate was because of the percieved
delay generating the codes but since it is quick it is not necessary for
the split.

Removed .active property since it will be available from the point the
class is initialised.

Also made generating the barcodes for jump() and next() in
BarcodeGenerator and showing the window (entering the mainloop) in
ChefShow optional but all default to enabled. This is to help speed up
testing to avoid the costly activities when some other feature is being tested.